Client Wallet Reorg
* update name from simplewallet to concealwallet
* update class to conceal_wallet
* create main.cpp strictly for main
* rename DepositHelper to ClientHelper
* add functions to clienthelper
* move transfer command into its own file
* remove calling wallet in main, just show flags not full wallet commands
* create Const header file
* add save_wallet function to client helper to save code duplication
*
